requirements.txt
datasets>=3.6,<4.0
numpy>=2.2,<3.0
PyYAML>=6.0,<7.0
ruff>=0.11,<0.12
torch>=2.10,<2.11
